Online editor: Move the property view code in a shared location
So that it can be re-used by the vscode extension
