+++ tools/online_editor/src/tsconfig.json
@@ -4,7 +4,7 @@
         "lib": ["dom", "es2020"],
         "types": ["json-schema", "vscode"]
     },
-    "include": ["./*.ts"],
+    "include": ["./*.ts", "./shared/*.ts"],
     "references": [
         {
             "path": "worker"
